test: introduce parent testsuite format
The current structure for unit testing only allows for running a
test suite with nested test cases. This means all test cases for an
autotest must be in one suite, which is not ideal.
For example, in some cases we may want to run multiple lists of test
cases that each require different setup, so should be in separate suites.

The unit test suite struct is modified to hold either a pointer to a
list of test cases, or a pointer to a list of sub-testsuites.
A bool value is also introduced here to indicate whether the testsuite
is a parent suite or not.

Signed-off-by: Ciara Power <ciara.power@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>