test: make hugepage check more robust under Linux
The hugepage test really needs to check multiple things on Linux:

1. Are hugepages reserved in the system?

2. Is the hugepage mountpoint available so that we can allocate them?

3. Do we have permissions to write into the hugepage mountpoint?

The existing hugepage check only verifies the first.  On some setups,
a non-root user won't have access to the mountpoint for hugepages to
be allocated and that needs to be reflected in the test as well.  Add
such checks for Linux OS to give a more check when running test suites.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>