mem: fix cleanup on Windows
rte_mem_unmap() called from rte_eal_memory_detach() fails on Windows
with message "EAL: Could not unmap memory: No error". This is because
on Windows memory is not allocated using mapping. Confusing "No error"
is caused by using errno instead of rte_errno set by rte_mem_unmap().

Multi-process is not supported on Windows and --in-memory is forced,
so detaching memory is not needed on cleanup. Bypass the function
in this case. Fix error handling to produce proper log message.

Fixes: dfbc61a2f9a6 ("mem: detach memsegs on cleanup")
Cc: Anatoly Burakov <anatoly.burakov@intel.com>

Reported-by: Jie Zhou <jizh@microsoft.com>
Suggested-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>