power: do not skip saving original pstate governor
Currently, when we set the pstate governor to "performance", we check if
it is already set to this value, and if it is, we skip setting it.

However, we never save this value anywhere, so that next time we come
back and request the governor to be set to its original value, the
original value is empty.

Fix it by saving the original pstate governor first. While we're at it,
replace `strlcpy` with `strscpy`.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>