net/i40e: fix mbuf fast free wrong check
If no offload flags are set, then 0 == 0 & <anything> is true.
This means that if DEV_TX_OFFLOAD_MBUF_FAST_FREE is not
set (and no other flags are set) then this check will still
enable `tx_simple_allowed`.

This patch adds the 'offloads' flag not equal zero to avoid this
scenario.

Bugzilla ID: 665
Fixes: 399421100e08 ("net/i40e: fix missing mbuf fast free offload")

Signed-off-by: Murphy Yang <murphyx.yang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>