examples/l2fwd-crypto: remove key size validation
In the example application the key can be provided by the user or
generated randomly by the example application.

Then a validation is done in order to check if the key size is
supported in the algorithm capabilities.

As a result of patch [1] application should support the case where a
user supplies a wrapped key.

In order to align with this patch and be able to use the app with
wrapped keys, app should remove the validation of key size in the
application and rely on a PMD key size validation.

The validation is removed in case the key is provided by by user, and
kept in case the key should be generated by the application.

[1] https://www.mail-archive.com/dev@dpdk.org/msg201281.html

Signed-off-by: Shiri Kuzin <shirik@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>