meson: remove unnecessary explicit link to libpcap
libpcap is already found and registered as a dependency by meson, and
the dependency is already correctly used in librte_port. This line is
just unnecessary.

The pcap PMD and the librte_port both declare their dependency to
libpcap with a line "ext_deps += pcap_dep".
"pcap_dep" is declared using the dependency() directive in
config/meson.build meson automatically adds this dependency to the
pkg-config file in the private section for static builds.

Signed-off-by: Gabriel Ganne <gabriel.ganne@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>